refactor: context types
