+++ src/core_modules/capture-core/metaDataStoreLoaders/context/context.js
@@ -1,39 +1,7 @@
 // @flow
-import { StorageController } from 'capture-core-utils/storage';
-import type { QueryApiFn } from '../loader.types';
+import type { ContextInput } from './context.types';
 
 let context;
-
-type StoreNames = {
-    CONSTANTS: string,
-    ORGANISATION_UNIT_LEVELS: string,
-    RELATIONSHIP_TYPES: string,
-    TRACKED_ENTITY_TYPES: string,
-    PROGRAMS: string,
-    PROGRAM_RULES: string,
-    PROGRAM_RULES_VARIABLES: string,
-    PROGRAM_INDICATORS: string,
-    TRACKED_ENTITY_ATTRIBUTES: string,
-    OPTION_SETS: string,
-    CATEGORIES: string,
-    CATEGORY_OPTIONS_BY_CATEGORY: string,
-    CATEGORY_OPTIONS: string,
-    REDUX_PERSIST: string,
-};
-
-type ParentStoreNames = {
-    USER_CACHES: string,
-    STATUS: string,
-    SYSTEM_SETTINGS: string,
-};
-
-type Input = {
-    onQueryApi: QueryApiFn,
-    storageController: StorageController,
-    parentStorageController: StorageController,
-    storeNames: StoreNames,
-    parentStoreNames: ParentStoreNames,
-};
 
 export const provideContext = async ({
     onQueryApi,
@@ -41,7 +9,7 @@ export const provideContext = async ({
     parentStorageController,
     storeNames,
     parentStoreNames,
-}: Input,
+}: ContextInput,
     callback: Function,
 ) => {
     context = {
